chore: adjust code position
