+++ extensions/ide/vscode/devbox/src/utils/file.ts
@@ -1,5 +1,7 @@
-import { execa } from 'execa'
 import * as os from 'os'
+import path from 'path'
+import * as fs from 'fs'
+import { execa } from 'execa'
 
 // File access permission modification
 export const ensureFileAccessPermission = async (path: string) => {
@@ -10,5 +12,16 @@ export const ensureFileAccessPermission = async (path: string) => {
     await execa('icacls', [path, '/remove:g', 'everyone'])
   } else {
     await execa('chmod', ['600', path])
+  }
+}
+
+export function ensureFileExists(filePath: string, parentDir: string) {
+  if (!fs.existsSync(filePath)) {
+    fs.mkdirSync(path.resolve(os.homedir(), parentDir), {
+      recursive: true,
+    })
+    fs.writeFileSync(filePath, '', 'utf8')
+    // .ssh/config authority
+    ensureFileAccessPermission(filePath)
   }
 }
